chore: Implement backend request (#110)
This Pull Request focusses on making a request to OpenZaak with changes
to the structure of the Add-In. In this PR there is a working request to
fetch a singular zaak from OpenZaak. the biggest changes are in
`ZaakService` and `app.ts`. There are also a lot of file creations for
more readable code structure and functionality.

Further measures have been taken to prevent the use of any in code with
some exceptions that have to be spoken about and handled in the future.

Solves PZ-7106