feature: setup frontend and backend locations (#170)
The manifest file is updated so that references to localhost points to
the environment frontend location.

The useHttps.ts file is updated so that it connects to the environment
backend location.

A test script has been added for manual verification that the
replacement script does what it is supposed to.

Solves PZ-7400