chore: updated zaaktypes in OpenZaak (#502)
This PR updates the zaaktypes in the docker setup of OpenZaak. It adds
the current test zaaktypes which we have in ZAC. These can be used for
testing in the office-add-in.

The `docker-compose.yml `has been changed in order to ensure that the
correct database filling files are used.

The config.sql file has been updated to ensure the correct secret key is
set

Solves PZ-9282