feat: Refactor of graph token and log env values (#560)
This PR Refactors the retrieving of the acces token for graph api acces
through MSAL. Local detection is also removed because MSAL must now work
in the local & test environment. There is an added log to see what the
frontend env variables are on the test environment which will be used
for debugging and research purposes.

Solves PZ-9590