feat: added MSAL to deployment (#561)
This PR add the required MSAL config to the provisioning so that Graph
can be accessed on the test environment. in this PR the following has
happened

- The env.frontend file has been merged with the root .env file. This
was decided after a discussion with @niorg. A small change therefore was
needed in the office-add-in/webpack.config.js.
- Removed .env.frontend from the gitignore
- Added the MSAL fields in the charts deployment. This has already been
prepared in the
[provisioning](https://github.com/infonl/dimpact-provisioning)
repository

Solves PZ-9591

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>