feat: put user info in JWT to open zaak (#610)
This PR implements the logic that the current logged in users `name` and
`preferred_username` is sent with the requests to open zaak. Changes in
the frontend and backend have been made for this.

Because tokens are now required the TokenError component has been
implemented to be used in Word & Excel environments

Tests have been altered to make use of the new POST logic

Solves PZ-9294 & PZ-9436